put PipelineExecutor into base library
